+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 ```
 func pushWithTemplate(templateName: String, templateValues: [String : Any], responseHandler: ((MessageStatus) -> Void)!, errorHandler: ((Fault) -> Void)!)
 ```
-* added method for channel:
+* added Channel function:
 ```
 func sendCommand(commandType: String, data: Any?, responseHandler: (() -> Void)!, errorHandler: ((Fault) -> Void)!)
 ```